Win screen and Crates
+Updated the win screen to include score and so that it looks like all the other pages

+Updated death screen to include score

+Opening the chest now leads you to the win screen.

+Implemented Crates
(open them to get matches, a banana, or nothing)

+Added a new function to rooms
(used for crates)

+Added more room descriptions
(The rooms are supposed to be divided into 4 different "zones". Stone, Dark/Foggy stone, steel, fancy. The idea is that this is supposed to make it easier to tell where you are. The rooms also have slightly different descriptions from one another to differentiate rooms in a single area.)

=Moved code from previous push into console servlet 2(room discovery, end score, end gem check)

-Removed code related to the NPC in PlayerModel.
(Honestly at this point I don't think we're getting the NPC so I cut the code I added for it.)

Will probably work on Hardy Mode next push.